chnaged a few ints to doubles to get better means

+++ src/squadMay21/stats/StandardDeviation.java
@@ -5,8 +5,7 @@ public class StandardDeviation {
 	
 	public static double getStandardDeviation(int arr[]) {
 		
-		//int avg = Mean.getMean(arr);
-		int vari = Variance.getVariance(arr);
-		return Math.sqrt((double)vari);
+		double vari = Variance.getVariance(arr);
+		return Math.sqrt(vari);
 	}
 }
